pyproject.toml
[tool.poetry]
name = "x402"
version = "0.1.0"
description = ""
authors = ["devcode8 <devccode8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastmcp = "^2.10.0"
httpx = "^0.27.0"
web3 = "^6.15.0"
eth-account = "^0.10.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0"
fastapi = "^0.104.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
